Updated neural net to fix new implementation description

+++ prometheus-ai/src/main/java/interfaces/SensorInput.java
@@ -1,5 +1,5 @@
 package interfaces;
 
 public interface SensorInput {
-	void receiveDataStream(int nnID, int nnStruct, int data[]);
+	void receiveDataStream(int nnID, int nnStruct, double data[]);
 }
